+++ clearing_current_day/HistoryOnHold.h
@@ -34,7 +34,7 @@ private:
 	virtual bool Get() override;
 	virtual bool IsExistData() override;
 	virtual bool CheckInsert(int _id) override;
-	virtual int Count() override;
+	virtual size_t Count() override;
 };
