+++ clearing_current_day/HistorySms.h
@@ -42,7 +42,7 @@ private:
 	virtual bool Get() override;
 	virtual bool IsExistData() override;
 	virtual bool CheckInsert(int _id) override;
-	virtual int Count() override;
+	virtual size_t Count() override;
 };
